requirements.txt
dj-database-url==0.5.0
Django==3.2.1
django-cors-headers==3.2.1
django-crispy-forms==1.9.2
djangorestframework==3.13
djangorestframework-simplejwt==4
django-htmx==1.7.0
Faker==0.9.1
gunicorn==20.0.4
Pillow==9.4.0
psycopg2-binary==2.9.5
six==1.15.0
urllib3==1.26.9
whitenoise==5.2.0
coverage==6.4.3
python-dotenv==0.21.0